+++ gems/factrey/spec/factrey/blueprint/type_spec.rb
@@ -31,4 +31,21 @@ RSpec.describe Factrey::Blueprint::Type do
       end
     end
   end
+
+  describe "#create_object" do
+    subject { type.create_object(context, *args, **kwargs) }
+
+    let(:type) do
+      described_class.new(:type) do |type, context, *args, **kwargs|
+        [context[:build_strategy], type.name, *args, kwargs]
+      end
+    end
+    let(:context) { { build_strategy: :build } }
+    let(:args) { [12, "34"] }
+    let(:kwargs) { { foo: :bar } }
+
+    it "calls the factory" do
+      expect(subject).to eq [:build, :type, 12, "34", { foo: :bar }]
+    end
+  end
 end