+++ gems/factrey/spec/factrey/proxy_spec.rb
@@ -3,18 +3,19 @@
 RSpec.describe Factrey::Proxy do
   let(:test_class) do
     Class.new do
-      def foo(name = nil, *args, **kwargs, &block)
-        return Factrey::Proxy.new(self, __method__) unless name
+      def foo(prefix, name = nil, *args, **kwargs, &block)
+        return Factrey::Proxy.new(self, __method__, prefix) unless name
 
-        [name, args, kwargs, block]
+        [prefix, name, args, kwargs, block]
       end
     end
   end
 
   it "forwards method calls to the receiver" do
-    expect(test_class.new.foo.bar(12, abc: "def") { 345 }).to match [
+    expect(test_class.new.foo(12).bar(34, abc: "def") { 345 }).to match [
+      12,
       :bar,
-      [12],
+      [34],
       { abc: "def" },
       have_attributes(call: 345),
     ]